[analytics] move ClientAnalytics into package @shopify/blockchain-components-analytics
