+++ packages/blockchain-components/src/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Utils
+ */
+export * from './utils/analytics/ClientAnalytics';
+
+/**
+ * Tests
+ */
+export {AnalyticsListenerTestHelper} from './tests/helpers/ClientAnalytics';
