+++ packages/blockchain-components/src/utils/analytics/ClientAnalytics/const.ts
@@ -1,3 +1,4 @@
 export const eventNames = {
   TOKENGATE_COMPONENT_RENDERED: 'TOKENGATE_COMPONENT_RENDERED',
+  CONNECT_BUTTON_CLICKED: 'CONNECT_BUTTON_CLICKED',
 };
