+++ packages/blockchain-components/src/utils/analytics/ClientAnalytics/index.ts
@@ -1,0 +1,2 @@
+export {eventNames} from './const';
+export {ClientAnalytics} from './utils';
