+++ packages/shared/src/components/Button/Button.test.tsx
@@ -1,7 +1,7 @@
+import {AnalyticsListenerTestHelper} from '@shopify/blockchain-components';
 import {fireEvent} from '@testing-library/react';
 import {vi} from 'vitest';
 
-import {AnalyticsListenerTestHelper} from '../../tests/helpers/ClientAnalytics';
 import {renderWithContext} from '../../tests/utils';
 
 import {Button} from './Button';
