+++ packages/shared/src/index.ts
@@ -50,4 +50,3 @@ export type {Theme, ThemeProps} from './types/theme';
 export {formatWalletAddress} from './utils/formatters';
 export {deepMerge} from './utils/deepMerge';
 export {addDays} from './utils/date';
-export {ClientAnalytics} from './utils/analytics';
