--- packages/shared/src/utils/analytics/index.ts
@@ -1,1 +1,0 @@
-export {ClientAnalytics} from './ClientAnalytics';
