+++ packages/tokengate/src/index.ts
@@ -21,8 +21,3 @@ export {Dawn, Default};
  */
 export type {Theme};
 export * from './types';
-
-/**
- * ClientAnalytics
- */
-export {ClientAnalytics} from 'shared';
